Simplify Value Aggregation for Group/Well Reports
In particular, switch to get_group_var() and get_well_var() instead
of the lower-level 'get()' function as the way to retrieve summary
values and replace lambdas called once with the actual logic.

Finally, switch to using named indices into the 'values' vector to
make the expressions (slightly) more readable.