Make Zoltan's Hyperedge Size Threshold Runtime Controllable
This commit introduces a new hidden runtime parameter,

    ZoltanPhgEdgeSizeThreshold (--zoltan-phg-edge-size-threshold,
    double, default value = 0.35)

which maps to Zoltan's low-level control parameter

    PHG_EDGE_SIZE_THRESHOLD

that controls which hypergraph edges to omit/discard.  We add a new
parameter to 'setupZoltanParameters()' and thread the command line
parameter through as an argument to this function.

This is to enable runtime experimentation with this aspect of the
partitioning algorithm, but end-users should typically not alter the
default value without good reason.